pyproject.toml
[tool.poetry]
name = "prject1"
version = "0.1.0"
description = ""
authors = ["lavrik91 <lavrik2837@icloud.com>"]


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
asyncpg = "^0.29.0"
psycopg = "^3.1.13"
pydantic-settings = "^2.1.0"
psycopg-binary = "^3.1.13"
pytest = "^7.4.3"
pytest-dotenv = "^0.5.2"
psycopg2 = "^2.9.9"
pytest-mock = "^3.12.0"
mock-alchemy = "^0.2.6"
pika = "^1.3.2"
httpx = "^0.25.2"
pytest-asyncio = "^0.23.2"
gunicorn = "^21.2.0"
flower = "^2.0.1"
psycopg2-binary = "^2.9.9"
sentry-sdk = "^1.39.1"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.3"}
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
radis = "^0.14"
fastapi-pagination = "^0.12.21"
rabbitmq = "^0.2.0"
loguru = "^0.7.2"
fastapi-logging = "^0.1.1"
celery = "^5.3.6"
aio-celery = "^0.10.0"
python-json-logger = "^2.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = [
    ".","src",
]
asyncio_mode = "auto"